requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.3.0
attrs==26.1.0
babel==2.18.0
beautifulsoup4==4.14.3
bitsandbytes==0.49.2
bleach==6.3.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
comm==0.2.3
cuda-bindings==13.0.3
cuda-pathfinder==1.2.2
datasets==4.8.3
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.1
einops==0.8.2
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.25.2
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
gitpython==3.1.46
h11==0.16.0
hf-xet==1.4.2
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.7.2
idna==3.11
imageio==2.37.3
ipykernel==7.2.0
ipython==9.11.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.3
json5==0.13.0
jsonpointer==3.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-client==8.8.0
jupyter-console==6.6.3
jupyter-core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.4
jupyterlab==4.5.6
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
jupyterlab-widgets==3.0.16
lark==1.3.1
lazy-loader==0.5
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.19
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
ninja==1.13.0
notebook==7.5.5
notebook-shim==0.2.4
numpy==2.3.5
nvidia-cublas==13.1.0.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-cu13==9.15.1.9
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver==12.0.4.66
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse==12.6.3.3
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu12==2.27.5
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu12==3.4.5
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
packaging==26.0
pandocfilters==1.5.1
parso==0.8.6
peft==0.18.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.9.4
prometheus-client==0.24.1
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
ptyprocess==0.7.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
pycparser==3.0
pydantic==2.12.5
pyarrow==22.0.0
pygments==2.19.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytorch-metric-learning @ git+https://github.com/KevinMusgrave/pytorch-metric-learning.git@c8350998ebc8aacf2c45de50e2556bc854cc0361
pyyaml==6.0.3
pyzmq==27.1.0
qwen-vl-utils==0.0.8
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
send2trash==2.1.0
sentencepiece==0.2.1
sentry-sdk==2.55.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
stack-data==0.6.3
sympy==1.14.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2026.3.3
timm==1.0.25
tinycss2==1.4.0
tokenizers==0.22.2
torch==2.10.0+cu130
torchvision==0.25.0+cu130
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.3.0
triton==3.6.0
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==4.0.15
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
